ssize_t as a return type of write/read handlers

+++ src/dummy_read_handler.hpp
@@ -15,8 +15,8 @@ public:
   DummyReadHandler(DummyReadHandler &&) = default;
   DummyReadHandler &operator=(DummyReadHandler &&) = default;
 
-  int handle(std::span<std::byte> buf,
-             __off64_t offset [[maybe_unused]]) noexcept override {
+  ssize_t handle(std::span<std::byte> buf,
+                 __off64_t offset [[maybe_unused]]) noexcept override {
     return buf.size();
   }
 };
