+++ src/read_handler_interface.hpp
@@ -2,12 +2,14 @@
 
 #include <cstddef>
 
-#include <linux/types.h>
+#include <bits/types.h>
+#include <sys/types.h>
 
 #include <span>
 
 #include "handler_interface.hpp"
 
 namespace cfq {
-using IReadHandler = IHandler<int(std::span<std::byte>, __off64_t) noexcept>;
+using IReadHandler =
+    IHandler<ssize_t(std::span<std::byte>, __off64_t) noexcept>;
 } // namespace cfq
